Move checker back to infra
