+++ doc/developers/python.rst
@@ -9,22 +9,3 @@ Client
 
 .. autoclass:: ccf.clients.Response
     :members:
-
-
-.. autoclass:: ccf.checker.Checker
-    :members: __call__
-
-Proposal Generator
-------------------
-
-.. autoclass:: ccf.proposal_generator.ProposalGenerator
-    :members:
-
-Ledger
-------
-
-.. autoclass:: ccf.ledger.Ledger
-    :members:
-
-
-
