+++ samples/apps/txregulator/tests/txregulatorclient.py
@@ -2,7 +2,7 @@
 # Licensed under the Apache 2.0 License.
 import infra.e2e_args
 import infra.network
-import ccf.checker
+import infra.checker
 import ccf.proposal_generator
 
 import logging
@@ -41,7 +41,7 @@ def run(args):
     with infra.network.network(
         hosts, args.binary_dir, args.debug_nodes, args.perf_nodes, pdb=args.pdb
     ) as network:
-        check = ccf.checker.Checker()
+        check = infra.checker.Checker()
         network.start_and_join(args)
         primary, others = network.find_nodes()
 
@@ -107,7 +107,7 @@ def run(args):
 
         # As permissioned manager, register regulator and banks
         with primary.client() as mc:
-            check_commit = ccf.checker.Checker(mc)
+            check_commit = infra.checker.Checker(mc)
 
             with primary.client(f"user{manager.name}") as c:
                 check(
