+++ tests/e2e_scenarios.py
@@ -7,7 +7,7 @@ import random
 import infra.network
 import infra.proc
 import infra.e2e_args
-import ccf.checker
+import infra.checker
 
 from loguru import logger as LOG
 
@@ -36,8 +36,8 @@ def run(args):
 
         with primary.client() as mc:
 
-            check = ccf.checker.Checker()
-            check_commit = ccf.checker.Checker(mc)
+            check = infra.checker.Checker()
+            check_commit = infra.checker.Checker(mc)
 
             for connection in scenario["connections"]:
                 with (
