+++ tests/infra/logging_app.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the Apache 2.0 License.
 
-import ccf.checker
+import infra.checker
 import ccf.clients
 import suite.test_requirements as reqs
 import time
@@ -89,8 +89,8 @@ class LoggingTxs:
     ):
         LOG.success(f"Applying {number_txs} logging txs to node {remote_node.node_id}")
         with remote_node.client() as mc:
-            check_commit = ccf.checker.Checker(mc)
-            check_commit_n = ccf.checker.Checker(mc, self.notifications_queue)
+            check_commit = infra.checker.Checker(mc)
+            check_commit_n = infra.checker.Checker(mc, self.notifications_queue)
 
             with remote_node.client(self.user) as uc:
                 for _ in range(number_txs):
@@ -166,7 +166,7 @@ class LoggingTxs:
                     LOG.warning("User frontend is not yet opened")
                     time.sleep(0.1)
                 else:
-                    check = ccf.checker.Checker(uc)
+                    check = infra.checker.Checker(uc)
                     check(
                         rep, result={"msg": txs[idx]},
                     )
