+++ tests/infra/member.py
@@ -7,7 +7,7 @@ import infra.node
 import infra.proposal
 import infra.crypto
 import ccf.clients
-import ccf.checker
+import infra.checker
 import http
 import os
 import base64
@@ -103,7 +103,7 @@ class Member:
                 # can only commit after it has successfully joined and caught up.
                 # Given that the retry timer on join RPC is 4 seconds, anything less is very
                 # likely to time out!
-                ccf.checker.wait_for_global_commit(mc, r.seqno, r.view, timeout=6)
+                ccf.commit.wait_for_commit(mc, r.seqno, r.view, timeout=6)
 
         return r
