+++ tests/receipts.py
@@ -6,7 +6,7 @@ import infra.notification
 import infra.net
 import suite.test_requirements as reqs
 import infra.e2e_args
-import ccf.checker
+import infra.checker
 
 from loguru import logger as LOG
 
@@ -18,8 +18,8 @@ def test(network, args, notifications_queue=None):
     primary, _ = network.find_primary_and_any_backup()
 
     with primary.client() as mc:
-        check_commit = ccf.checker.Checker(mc, notifications_queue)
-        check = ccf.checker.Checker()
+        check_commit = infra.checker.Checker(mc, notifications_queue)
+        check = infra.checker.Checker()
 
         msg = "Hello world"
